Extract Turbo actions in a custom `turbo.ts file
