+++ README.md
@@ -102,15 +102,12 @@ Checkout the instructions in the [`turbo_power-rails`](https://github.com/marcor
 
 ### Browser Actions
 
-* `turbo_stream.redirect_to(url, turbo_action = nil, **attributes)`
 * `turbo_stream.reload(**attributes)`
 * `turbo_stream.scroll_into_view(target, inline = "nearest")`
 * `turbo_stream.set_cookie(cookie, **attributes)`
 * `turbo_stream.set_cookie_item(key, value, **attributes)`
 * `turbo_stream.set_focus(target, **attributes)`
 * `turbo_stream.set_title(title, **attributes)`
-* `turbo_stream.turbo_clear_cache()`
-
 
 ### Browser History Actions
 
@@ -133,6 +130,11 @@ Checkout the instructions in the [`turbo_power-rails`](https://github.com/marcor
 
 * `turbo_stream.turbo_frame_reload(frame_id)`
 * `turbo_stream.turbo_frame_set_src(frame_id, src)`
+
+## Turbo Actions
+
+* `turbo_stream.redirect_to(url, turbo_action = nil, **attributes)`
+* `turbo_stream.turbo_clear_cache()`
 
 ## Previous Art
