+++ src/actions.ts
@@ -8,6 +8,7 @@ import * as Events from "./actions/events"
 import * as History from "./actions/history"
 import * as Storage from "./actions/storage"
 import * as TurboFrame from "./actions/turbo_frame"
+import * as Turbo from "./actions/turbo"
 
 export * from "./actions/attributes"
 export * from "./actions/browser"
@@ -17,6 +18,7 @@ export * from "./actions/events"
 export * from "./actions/history"
 export * from "./actions/storage"
 export * from "./actions/turbo_frame"
+export * from "./actions/turbo"
 
 export function register(streamActions: TurboStreamActions) {
   Attributes.registerAttributesActions(streamActions)
@@ -27,4 +29,5 @@ export function register(streamActions: TurboStreamActions) {
   History.registerHistoryActions(streamActions)
   Storage.registerStorageActions(streamActions)
   TurboFrame.registerTurboFrameActions(streamActions)
+  Turbo.registerTurboActions(streamActions)
 }
